refactor: rename components, place card into separate component
