+++ react-rss/src/components/DetailedCard/detailedCard.scss
