+++ react-rss/src/components/DetailedHeader/detailedHeader.scss
