+++ react-rss/src/components/DetailedProps/detailedProps.scss
