+++ react-rss/src/components/SearchResults/SearchResults.tsx
@@ -1,11 +1,10 @@
 import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
+import Card from '../Card/Card';
 import Loader from '../Loader/Loader';
 import Pagination from '../Pagination/Pagination';
 import NotFound from '../NotFound/NotFound';
 import { Character, ResultData } from '../../utilities/types';
-
-import location from '../../assets/location.svg';
 
 import './searchResults.scss';
 
@@ -22,7 +21,7 @@ export default function SearchResults({
   resultData,
   success,
 }: SearchResultsProps): JSX.Element {
-  const [searchParams, setSearchParams] = useSearchParams();
+  const [searchParams] = useSearchParams();
 
   useEffect(() => {
     loadPage(Number(searchParams.get('page')), false);
@@ -41,33 +40,7 @@ export default function SearchResults({
     <div className="search-results">
       <ul className="search-results__list">
         {resultData?.results.map((character: Character) => (
-          <li
-            className="character"
-            key={character.id}
-            onClick={() => {
-              searchParams.set('id', `${character.id}`);
-              setSearchParams(searchParams);
-            }}
-          >
-            <img
-              className="character__img"
-              src={character.image}
-              alt={character.name}
-            />
-            <div
-              className={`character__status character__status_${character.status.toLowerCase()}`}
-            >
-              {character.status}
-            </div>
-            <div className="character__info">
-              <h3 className="character__name">{character.name}</h3>
-              <p className="character__species">{character.species}</p>
-              <p className="character__location">
-                <img className="location-icon" src={location} alt="location" />
-                {character.location.name}
-              </p>
-            </div>
-          </li>
+          <Card key={character.id} character={character} />
         ))}
       </ul>
       <Pagination pageQuantity={resultData?.info.pages} />
